requirements.txt
fastapi
uvicorn
apscheduler
requests
mysql-connector-python
pathlib
argparse